% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{predict.rjMCMC}
\alias{predict.rjMCMC}
\title{Predict from Reversible Jump MCMC (rjMCMC) Model Results}
\usage{
\method{predict}{rjMCMC}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class rjMCMC containing the output from the \code{rjMCMC} procedure, including posterior samples and fitted splines.}

\item{newdata}{A data frame for which predictions are to be made. If NA, the original fitted data is used.}

\item{...}{Additional arguments to be passed to other methods or functions.}
}
\value{
A matrix of predicted values.
}
\description{
This function generates posterior predictions from an rjMCMC model based on the provided data.
It combines the fixed effects, spline terms, and binary parameters, and includes residual variance
in the predictions.
}
\examples{
\donttest{
# Example dataset
data("simulated_data")

candsplinevars <- c("X_1")
candbinaryvars <- paste0("Z_", 1:5)
candinter <- c(candsplinevars, candbinaryvars)

mcmc_specs <- list(B = 2000, burnin = 1000, thin = 1, chains = 2, sigma_v = 0.1, bma = TRUE)
prior_params <- list(lambda_1 = 0.1, lambda_2 = 1, a_0 = 0.01, b_0 = 0.01,
                  degree = 3, k_max = 9, w = 1, sigma_B = sqrt(20))

results <- rjMCMC(simulated_data, candsplinevars, candbinaryvars, candinter,
                  mcmc_specs, prior_params)
newdata = data.frame(Z_1 = 1, Z_2 = 1, Z_3 = 1, Z_4 = 1, Z_5 = 1,
                   trt = 1, X_1 = seq(0,1,by=0.01))
predict(results)
predict(results,newdata)
}
}
